added confirmation dialog for deleting view points + progress spinner

+++ src/components/PointsTable.js
@@ -1,28 +1,39 @@
 import React, { useState } from 'react'
 import { makeStyles } from '@material-ui/core/styles'
+import { useConfirm } from 'material-ui-confirm'
 import {
-  Table, TableBody, TableRow, TableCell, TableFooter,
-  TablePagination, TableContainer, Paper, IconButton
+  Table, TableBody, TableRow, TableCell, TableFooter, TablePagination,
+  TableContainer, Paper, IconButton, CircularProgress
 } from '@material-ui/core'
 import DeleteForeverIcon from '@material-ui/icons/DeleteForever'
-import EditForm from 'components/EditForm'
-import Pagination from 'components/Pagination'
-import PointDialog from 'components/PointDialog'
 import { POINTS_FILENAME } from 'assets/constants'
+import PointDialog from 'components/PointDialog'
+import Pagination from 'components/Pagination'
+import EditForm from 'components/EditForm'
 import _ from 'lodash'
 
 const useStyles = makeStyles({
   table: {
-    minWidth: 300,
+    minWidth: 260,
+  },
+  buttonProgress: {
+    color: 'red',
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    marginTop: -12,
+    marginLeft: -12,
   },
 })
 
 const PointsTable = (props) => {
   const classes = useStyles()
+  const confirm = useConfirm()
   const { hub, points, isOwner, userSession, updatePoints } = props
 
   const [page, setPage] = useState(0)
   const [rowsPerPage, setRowsPerPage] = useState(5)
+  const [loading, setLoading] = useState(false);
 
   const emptyRows = rowsPerPage - Math.min(rowsPerPage, points.length - page * rowsPerPage)
 
@@ -36,16 +47,27 @@ const PointsTable = (props) => {
   const handleDelete = async (pointId) => {
     const filteredPoints = _.filter(points, (point) => point.id !== pointId)
 
-    await userSession.putFile(
-      POINTS_FILENAME,
-      JSON.stringify(filteredPoints),
-      { encrypt: false }
-    )
-    .catch(error => { console.log(error.message) })
+    await confirm({
+      title: 'Are you sure you want to delete your view point?',
+      description: 'This action is permanent!',
+      confirmationButtonProps: { color: 'secondary' }
+    }).then(() => {
+      if (!loading) {
+        setLoading(true)
 
-    await userSession.deleteFile(`point-${pointId}.json`)
-      .then(() => { updatePoints(filteredPoints) })
-      .catch(err => { console.log(err.message) })
+        userSession.putFile(
+          POINTS_FILENAME,
+          JSON.stringify(filteredPoints),
+          { encrypt: false }
+        ).catch(error => { console.log(error.message) })
+
+        userSession.deleteFile(`point-${pointId}.json`)
+          .catch(err => { console.log(err.message) })
+      }
+    }).then(() => {
+      updatePoints(filteredPoints)
+      setLoading(false)
+    }).catch((e) => { console.log(e) })
   }
 
   return (
@@ -56,10 +78,9 @@ const PointsTable = (props) => {
             ? points.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
             : points
           ).map((row) => (
-            <TableRow key={row.id ? row.id : '1'}>
+            <TableRow key={row.id}>
               <TableCell>
                 <PointDialog id={row.id} hub={hub} title={row.title} />
-                &nbsp; {row.title}
               </TableCell>
               <TableCell align="right">
                 {
@@ -78,6 +99,7 @@ const PointsTable = (props) => {
                       onClick={() => {handleDelete(row.id)}}
                     >
                       <DeleteForeverIcon fontSize="small" />
+                      {loading && <CircularProgress size={24} className={classes.buttonProgress} />}
                     </IconButton>
                   </>
                 }
@@ -95,7 +117,7 @@ const PointsTable = (props) => {
           <TableRow>
             <TablePagination
               rowsPerPageOptions={[5, 10, { label: 'All', value: -1 }]}
-              colSpan={3}
+              colSpan={2}
               count={points.length}
               rowsPerPage={rowsPerPage}
               page={page}
